pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "app-builder"
version = "0.1.0"
description = "API for building and deploying applications based on user requests"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "PyGithub>=2.1.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.4.2",
    "gitpython>=3.1.40",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[project.scripts]
app-builder = "app.main:app"
